consumer/requirements.txt
kafka-python>=1.4.7
pandas == 1.2.4
pyspark == 3.2.0